Refactor ListInput editing into generic class
